requirements.txt
flask
twilio
psycopg2-binary
openai
python-dotenv
gunicorn
